backend/requirements.txt
streamlit==1.24.0
plotly==5.15.0
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
requests==2.28.2
pandas==1.5.3
pillow==9.4.0
numpy==1.24.3
deepface==0.0.79
pydantic==1.10.7 
opencv-python==4.7.0.72
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
email-validator==2.0.0
